feat: replace eventos page with agenda page and update all references
- Transform /eventos page content into /agenda page
- /eventos now redirects to /agenda
- Update all nav links (mega-menu, footer, home, search, hero) to /agenda
- Update search page results and site pages to point to /agenda
